YeyoCogio

+++ views.py
@@ -1,7 +1,11 @@
 #coding=utf-8
 import os
 import webapp2
+from google.appengine.ext import webapp
+from google.appengine.api import users
 import jinja2
+
+import json
 
 from models import Comic, Entrega
 
@@ -9,6 +13,7 @@ from google.appengine.ext import ndb
 
 from google.appengine.ext import gql
 from google.appengine.api import images
+from google.appengine.api import urlfetch
 import datetime #esto lo vamos a usar para pasar de string a fecha.
 from google.appengine.api import users
 
@@ -40,14 +45,30 @@ class login(BaseHandler):
 class showComics(BaseHandler):
   
     def get(self):
+        user = users.get_current_user()
+
+        if user:
         
-        lista=Comic.query()
-        self.render_template('comics.html', {'listaComic': lista})
+            lista=Comic.query()
+            self.render_template('comics.html', {'listaComic': lista})
+        else:
+            self.redirect(users.create_login_url(self.request.uri)) 
+  
+  
 
 class AddComic(BaseHandler):
 
     def get(self):
-        self.render_template('newComic.html', {})
+        
+        user = users.get_current_user()
+
+        if user:
+        
+            self.render_template('newComic.html', {})
+        else:
+            self.redirect(users.create_login_url(self.request.uri))
+        
+        
         
     def post(self):
         imagenPortada=self.request.get('portadaComic')
@@ -70,14 +91,34 @@ class Image(webapp2.RequestHandler):
         else:
             self.response.out.write('No image')
 # [END image_handler]
+
+class ImagenEntrega(webapp2.RequestHandler):
+    def get(self):
+        Entrega_key = ndb.Key(urlsafe=self.request.get('img_id'))
+        comi = Entrega_key.get()
+        if comi.imagen:
+            self.response.headers['Content-Type'] = 'image/png'
+            self.response.out.write(comi.imagen)
+        else:
+            self.response.out.write('No image')
     
     
 class EditComic(BaseHandler):
 
     def get(self, comicID):
-        comic_id= int (comicID)
-        com=Comic.get_by_id(comic_id)
-        self.render_template('editarComic.html', {'comic': com})
+        
+        user = users.get_current_user()
+
+        if user:
+        
+            comic_id= int (comicID)
+            com=Comic.get_by_id(comic_id)
+            self.render_template('editarComic.html', {'comic': com})
+        else:
+            self.redirect(users.create_login_url(self.request.uri))
+        
+        
+        
         
     def post(self,comicID):
         comic_id= int (comicID)
@@ -91,40 +132,77 @@ class EditComic(BaseHandler):
 class DeleteComic(BaseHandler):
 
     def get(self, comicID):
-        print comicID
-        comic_id= int (comicID)
-        com=Comic.get_by_id(comic_id)
-        com.key.delete()
-        return webapp2.redirect('/')
+        
+        user = users.get_current_user()
 
+        if user:
+        
+            comic_id= int (comicID)
+            com=Comic.get_by_id(comic_id)
+            com.key.delete()
+            return webapp2.redirect('/')
+        else:
+            self.redirect(users.create_login_url(self.request.uri))
+        
 
 class BuscarFechaMayor(BaseHandler):
     def get(self):
-        fechaString=self.request.get('busquedaFechaMayor')
-        fechaDate=datetime.datetime.strptime(fechaString,'%Y-%m-%d')
-        q=Comic.gql("WHERE fechaCreacion >:fech",fech=fechaDate)
+        
+        user = users.get_current_user()
+
+        if user:
+        
+            fechaString=self.request.get('busquedaFechaMayor')
+            fechaDate=datetime.datetime.strptime(fechaString,'%Y-%m-%d')
+            q=Comic.gql("WHERE fechaCreacion >:fech",fech=fechaDate)
         #query=Comic.query("fechaCreacion >"+fecha)#strftime castea de string a tipo date.
-        print q
         #cos=Controller().listaFechaMayor(fecha)
-        self.render_template('comics.html', {'listaComic': q})
+            self.render_template('comics.html', {'listaComic': q})
+        else:
+            self.redirect(users.create_login_url(self.request.uri))
+            
+        
+        
+        
+        
 
 
 class showEntregasComic(BaseHandler):
     def get(self,comicID):
-        comic_id= int (comicID)
-        com=Comic.get_by_id(comic_id)
         
-        listaEntregas=Entrega.query(Entrega.idComic==com.key)
+        
+        user = users.get_current_user()
+
+        if user:
+        
+            comic_id= int (comicID)
+            com=Comic.get_by_id(comic_id)
+        
+            listaEntregas=Entrega.query(Entrega.idComic==com.key)
        
-        self.render_template('entregasComicPrueba.html', {'listaEntregas': listaEntregas,'comicID': comicID})
+            self.render_template('entregasComicPrueba.html', {'listaEntregas': listaEntregas,'comicID': comicID})
+        else:
+            self.redirect(users.create_login_url(self.request.uri))
+        
+        
+        
+        
 
 
 class AddEntrega(BaseHandler):
 
     def get(self, comicID):
-        comic_id= int (comicID)
-        com=Comic.get_by_id(comic_id)
-        self.render_template('newEntrega.html', {'comic':com})
+        
+        user = users.get_current_user()
+
+        if user:
+        
+            comic_id= int (comicID)
+            com=Comic.get_by_id(comic_id)
+            self.render_template('newEntrega.html', {'comic':com})
+        else:
+            self.redirect(users.create_login_url(self.request.uri))
+
         
     def post(self,comicID):
         comic_id= int (comicID)
@@ -144,18 +222,36 @@ class AddEntrega(BaseHandler):
 class DeleteEntrega(BaseHandler):
 
     def get(self, entregaID):
+        
+        
+        
+        user = users.get_current_user()
+
+        if user:
+            entrega_id= int (entregaID)
+            en=Entrega.get_by_id(entrega_id)
+            en.key.delete()
+            return webapp2.redirect('/')
+        else:
+            self.redirect(users.create_login_url(self.request.uri))
        
-        entrega_id= int (entregaID)
-        en=Entrega.get_by_id(entrega_id)
-        en.key.delete()
-        return webapp2.redirect('/')
+        
     
 class EditEntrega(BaseHandler):
 
     def get(self, entregaID):
-        entrega_id= int (entregaID)
-        en=Entrega.get_by_id(entrega_id)
-        self.render_template('editarEntrega.html', {'entrega': en})
+        
+        user = users.get_current_user()
+
+        if user:
+            entrega_id= int (entregaID)
+            en=Entrega.get_by_id(entrega_id)
+            self.render_template('editarEntrega.html', {'entrega': en})
+        else:
+            self.redirect(users.create_login_url(self.request.uri))
+        
+        
+        
         
     def post(self,entregaID):
         entrega_id= int (entregaID)
@@ -165,3 +261,63 @@ class EditEntrega(BaseHandler):
         en.put()
         
         return webapp2.redirect('/')  
+    
+class loginPason(webapp.RequestHandler):
+    
+    
+    def get(self):
+        user = users.get_current_user()
+
+        if user:
+            
+            self.redirect("/comics")
+        else:
+            self.redirect(users.create_login_url(self.request.uri)) 
+class logout(webapp.RequestHandler):
+    
+    
+    def get(self):
+        
+        self.redirect(users.create_logout_url("/"))             
+
+class flickr(BaseHandler):
+    def get(self):
+        
+        
+        user = users.get_current_user()
+
+        if user:
+            tag = self.request.get('nombreApi')
+        
+            clave = "b56fe120d5d60193cdff938b0af0ed91"
+        #secreto = '7941625e2330f77c'
+            url = 'https://api.flickr.com/services/rest/?method=flickr.photos.search&format=json&api_key='+clave+'&tags='+tag
+            r = urlfetch.fetch(url)
+            
+           
+            r = r.content[14:]
+            r = r[:-1]
+            respuesta = json.loads(r)
+            lista=  respuesta['photos']['photo']
+           
+            listaUrl = []
+            for i in range(len(lista)):
+                
+                secret = lista[i]['secret']
+                farm = lista[i]['farm']
+                server = lista[i]['server']  
+                id=lista[i]['id']
+                urlfoto = 'http://farm'+str(farm)+'.staticflickr.com/'+str(server)+'/'+str(id)+'_'+str(secret)+'_z.jpg'
+                listaUrl.append(urlfoto)
+            
+            
+            self.render_template('flickr.html', {'lista': listaUrl}) 
+        else:
+            self.redirect(users.create_login_url(self.request.uri))
+        
+        
+        
+        
+        
+            
+
